bombs on space ready
