+++ public_html/js/views/GameModules/gameObjects.js
@@ -12,6 +12,7 @@ define(function (require) {
         secondCharacter: null,
         objects: {}, // here we dump all links to obstacle index by id of object
         obstacles: [], // here we dump all our obstacles for raycaster
+        bombObj: null,
         getRealCoordinates: function (x, z) {
             return {
                 x: x * 64 - 992,
@@ -56,6 +57,19 @@ define(function (require) {
                 delete this.objects[id];
             }
         },
+        setBomb: function (id) {
+            var bomb = this.bombObj.clone();
+            bomb.position.set(this.firstCharacter.mesh.position.x, 2, this.firstCharacter.mesh.position.z);
+            var timerId = setInterval(function () {
+                bomb.scale.y *= 1.2;
+                bomb.scale.x *= 1.2;
+                bomb.scale.z *= 1.2;
+            }, 1000);
+            setTimeout(function () {
+                clearInterval(timerId);
+            }, 3000);
+            this.scene.add(bomb);
+        }
 };
     
     return objects;
